Roles semi-work

+++ src/main/java/ru/knowledgebase/dbmodule/repositories/ArticleRepository.java
@@ -9,7 +9,7 @@ import ru.knowledgebase.modelsmodule.Article;
 
 import java.util.List;
 
-public interface ArticleRepository extends CrudRepository<Article, Long> {
+public interface ArticleRepository extends CrudRepository<Article, Integer> {
     @Query("from Article")
     public List<Article> getAll() throws Exception;
 }
